backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-environ==0.11.2
django-cors-headers==4.4.0
Pillow==10.4.0
psycopg2-binary==2.9.9
openai==1.51.2

